requirements.txt
cohere
python-dotenv
pandas
pinecone
